ThinkCompress: pivot to CoT KV cache compression + full experiment results
Major direction change: from SpecQuant (spec decode KV quant) to ThinkCompress
(adaptive KV cache compression for chain-of-thought reasoning).

Core new module:
- src/thinkcompress.py (1602 lines): ThinkCompressor, ImportanceScorer,
  AdaptiveBitAllocator, CompressedThinkingCache. Tracks per-token attention
  importance during CoT thinking, assigns adaptive bit-widths, compresses
  low-importance thinking KV while preserving conclusion tokens at FP16.

Key experimental findings (2×H100 80GB):
- Qwen3-8B "17*23": 3346 thinking tokens (88%) vs 457 answer tokens (12%)
- Thinking KV dominates memory: 88-100% across all tested scenarios
- Qwen3-72B complex reasoning: thinking KV alone reaches 55GB (exceeds 1×H100)
- Qwen2.5-7B→14B spec decode: SpecQuant 4-bit 88.8% acceptance (baseline data)
- Microbenchmark: 35 configs across seq_len/methods/bits

Also includes:
- src/quantized_verifier.py: monkey-patched attention (MHA models)
- src/linear_attn_quantizer.py: GatedDeltaNet state quantization (Qwen3.5)
- src/baselines.py: RTN/KIVI/Absmax baselines
- Full experiment plan (refine-logs/EXPERIMENT_PLAN.md)
- 42 unit tests passing
- 4 benchmark result JSON files + logs

+++ src/turboquant_kv.py
@@ -261,7 +261,7 @@ class QuantizedKVCache:
             padded_dim = self.rotation.padded_dim
             scale = 1.0 / math.sqrt(padded_dim)
 
-        q_rotated = self.rotation.rotate(query)
+        q_rotated = self.rotation.rotate(query.float())
 
         k_rotated, v_rotated = self.get_rotated_kv(layer_idx)
